+++ src/FreeElephants/RestDaemon/Middleware/Collection/AbstractMiddlewareCollection.php
@@ -2,6 +2,8 @@
 
 namespace FreeElephants\RestDaemon\Middleware\Collection;
 
+use FreeElephants\RestDaemon\RestServer;
+
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
@@ -10,11 +12,22 @@ abstract class AbstractMiddlewareCollection extends \ArrayObject implements Midd
 
     protected $defaultBeforeMiddlewareMap;
 
-    public function __construct($additionalMiddleware = [])
+    /**
+     * @var RestServer
+     */
+    protected $server;
+
+    public function __construct(RestServer $restServer, $additionalMiddleware = [])
     {
+        $this->setServer($restServer);
         $middleware = array_merge($this->getDefaultMiddleware(), $additionalMiddleware);
         parent::__construct($middleware);
     }
 
     abstract protected function getDefaultMiddleware(): array;
+
+    public function setServer(RestServer $restServer)
+    {
+        $this->server = $restServer;
+    }
 }