+++ src/FreeElephants/RestDaemon/Middleware/Collection/MiddlewareCollectionInterface.php
@@ -2,6 +2,8 @@
 
 namespace FreeElephants\RestDaemon\Middleware\Collection;
 
+use FreeElephants\RestDaemon\RestServer;
+
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
@@ -10,4 +12,5 @@ interface MiddlewareCollectionInterface extends \ArrayAccess
 
     public function getArrayCopy();
 
+    public function setServer(RestServer $restServer);
 }