+++ src/FreeElephants/RestDaemon/Middleware/MiddlewareRole.php
@@ -13,4 +13,5 @@ class MiddlewareRole
     const BODY_PARSER = 'body-parser';
     const NO_CONTENT_STATUS_SETTER = 'no-content-status-setter';
     const CORS_HEADER_SETTER = 'CORS-header-setter';
+    const X_POWERED_HEADER_SETTER = 'x-powered-header-setter';
 }