+++ tests/unit/FreeElephants/RestDaemon/Middleware/Collection/DefaultBeforeMiddlewareCollectionTest.php
@@ -5,6 +5,7 @@ namespace FreeElephants\RestDaemon\Middleware\Collection;
 use FreeElephants\AbstractTestCase;
 use FreeElephants\RestDaemon\Middleware\ContentTypeSetter;
 use FreeElephants\RestDaemon\Middleware\MiddlewareRole;
+use FreeElephants\RestDaemon\RestServer;
 
 /**
  * @author samizdam <samizdam@inbox.ru>
@@ -14,7 +15,7 @@ class DefaultBeforeMiddlewareCollectionTest extends AbstractTestCase
 
     public function testDefaultMiddlewareInstances()
     {
-        $stack = new DefaultBeforeMiddlewareCollection();
+        $stack = new DefaultBeforeMiddlewareCollection($this->createMock(RestServer::class));
         self::assertInstanceOf(ContentTypeSetter::class, $stack->offsetGet(MiddlewareRole::CONTENT_TYPE_SETTER));
     }
 }